+++ src/accessory/helpers.ts
@@ -1,11 +1,13 @@
 import { PlatformAccessory } from 'homebridge';
 
 import { DummyAccessory } from './base.js';
+import { BlindAccessory } from './blind.js';
 import { DoorAccessory } from './door.js';
 import { LightbulbAccessory } from './lightbulb.js';
 import { LockAccessory } from './lock.js';
 import { OutletAccessory } from './outlet.js';
 import { SwitchAccessory } from './switch.js';
+import { WindowAccessory } from './window.js';
 
 import { strings } from '../i18n/i18n.js';
 
@@ -34,6 +36,10 @@ export function createAccessory(
     return new OutletAccessory(Service, Characteristic, accessory, config, log, persistPath, isGrouped);
   case Types.AccessoryType.Switch:
     return new SwitchAccessory(Service, Characteristic, accessory, config, log, persistPath, isGrouped);
+  case Types.AccessoryType.Window:
+    return new WindowAccessory(Service, Characteristic, accessory, config, log, persistPath, isGrouped);
+  case Types.AccessoryType.WindowCovering:
+    return new BlindAccessory(Service, Characteristic, accessory, config, log, persistPath, isGrouped);
   default:
     log.error(strings.startup.unsupportedType, `'${config.type}'`);
     return null;